# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nấc Thang API",
    "description": "Scoring and ranking for a group that competes. Every route except /health and /auth/discord needs a session token, and every route inside a workspace checks membership on the server.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token returned by POST /auth/discord"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness probe",
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/discord": {
      "post": {
        "summary": "Exchange a Discord authorization code for a session token",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "redirectUri"
                ],
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "redirectUri": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  }
                }
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Sign out this device",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "summary": "Sign out every device, revoking all sessions",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "The signed-in account and its workspaces",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/search": {
      "get": {
        "summary": "Search workspaces by name, ignoring Vietnamese diacritics",
        "tags": [
          "workspaces"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 60
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces": {
      "post": {
        "summary": "Create a workspace; the caller becomes its owner",
        "tags": [
          "workspaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "isPublic"
                ],
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/join": {
      "post": {
        "summary": "Join a public workspace, or ask the owner of a private one",
        "tags": [
          "workspaces"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/leave": {
      "post": {
        "summary": "Leave a workspace; points are kept for when you come back",
        "tags": [
          "workspaces"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/board": {
      "get": {
        "summary": "The leaderboard, the caller's own row and the scoring rules",
        "tags": [
          "workspaces"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}": {
      "patch": {
        "summary": "Rename a workspace or switch it between public and private (owner only)",
        "tags": [
          "workspaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/members": {
      "get": {
        "summary": "Everyone in the workspace, pending requests included (owner only)",
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/members/{userId}/approve": {
      "post": {
        "summary": "Approve a join request (owner only)",
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/members/{userId}": {
      "delete": {
        "summary": "Decline a request or remove a member (owner only)",
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/matches": {
      "get": {
        "summary": "The most recent matches in a workspace",
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Record a result and get the updated board back",
        "tags": [
          "matches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "opponentId",
                  "result"
                ],
                "additionalProperties": false,
                "properties": {
                  "opponentId": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "result": {
                    "type": "string",
                    "enum": [
                      "win",
                      "loss"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/matches/{matchId}": {
      "delete": {
        "summary": "Undo a match and give both players their points back (owner only)",
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "matchId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/workspaces/{id}/events": {
      "get": {
        "summary": "Server-sent events for one workspace; the payload only says what changed",
        "tags": [
          "events"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,9}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "Discord sign-in and session lifetime"
    },
    {
      "name": "workspaces",
      "description": "Groups, each with its own board"
    },
    {
      "name": "members",
      "description": "Approving and removing people, owner only"
    },
    {
      "name": "matches",
      "description": "Recording and undoing results"
    },
    {
      "name": "events",
      "description": "The realtime channel"
    }
  ]
}